requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-migrate==4.0.4
flask-wtf==1.0.1
werkzeug==2.2.3
openai ==0.27.8
python-dotenv==0.21.1
pillow==9.4.0
email-validator==2.0.0
gunicorn==20.1.0